requirements.txt
elasticsearch8
python-dotenv
sentence-transformers
tabulate
matplotlib
pandas
numpy
scipy
requests
urllib3